+++ zio-query/shared/src/main/scala/zio/query/DataSource.scala
@@ -40,6 +40,12 @@ trait DataSource[-R, -A] { self =>
   def runAll(requests: Iterable[Iterable[A]]): ZIO[R, Nothing, CompletedRequestMap]
 
   /**
+   * Syntax for adding aspects.
+   */
+  def @@[R1](aspect: DataSourceAspect[R, R1]): DataSource[R1, A] =
+    aspect(self)
+
+  /**
    * Returns a new data source that executes requests of type `B` using the
    * specified function to transform `B` requests into requests that this data
    * source can execute.
